r3149: Small tweak.
...

OK, actually, not so small. Removing the old patches section, and moving a few
last things from it into other sections. Also changing uses of macros from the
old model section to the new macros for the same things. Reordering a few other
paragraphs and sections, updating some old text, and adding up-to-date timing
information to the evaluation section.