r4314: Fixing mysterious files that should have been renamed/deleted.
Turns out that git-svnimport creates a list of files to add and remove in each
svn checkin, and then applies those lists afterward. But it deleted first and
then added, so files which were renamed as part of a directory rename and then
renamed inside the new directory would not have their original names deleted
from the new directory. Fixed by using a local version of git-svnimport that
applies these changes in the other order.